Refactor Inserter to dedicate component
